# server-go/docs/swagger.yaml
basePath: /api
definitions:
...
  models.AddFavoriteInput:
    properties:
      productId:
        type: integer
    required:
    - productId
    type: object
  models.AddToCartInput:
    properties:
      productId:
        type: integer
      quantity:
        type: integer
    required:
    - productId
    type: object
...
  models.Banner:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      imageUrl:
        type: string
      isActive:
        type: boolean
      link:
        type: string
      sortOrder:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.SiteConfig:
    properties:
      configKey:
        type: string
      configValue:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
    type: object
...
  models.UpdateCartInput:
    properties:
      quantity:
        type: integer
    required:
    - quantity
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@bycigar.com
    name: API Support
  description: BYCIGAR 电商后台 API 接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: BYCIGAR API
  version: "1.0"
paths:
...
  /admin/banners:
    post:
      consumes:
      - application/json
      description: 创建新的轮播图
      parameters:
      - description: 轮播图信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Banner'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Banner'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建轮播图
      tags:
      - admin-banners
  /admin/banners/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的轮播图
      parameters:
      - description: 轮播图ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除轮播图
      tags:
      - admin-banners
    put:
      consumes:
      - application/json
      description: 更新指定的轮播图
      parameters:
      - description: 轮播图ID
        in: path
        name: id
        required: true
        type: integer
      - description: 轮播图信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Banner'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Banner'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新轮播图
      tags:
      - admin-banners
...
  /admin/config/{key}:
    put:
      consumes:
      - application/json
      description: 更新指定的网站配置项
      parameters:
      - description: 配置键名
        in: path
        name: key
        required: true
        type: string
      - description: 配置值
        in: body
        name: input
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SiteConfig'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新网站配置
      tags:
      - admin-config
...
  /cart:
    get:
      consumes:
      - application/json
      description: 获取当前用户的购物车内容
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取购物车
      tags:
      - cart
    post:
      consumes:
      - application/json
      description: 添加商品到当前用户的购物车
      parameters:
      - description: 商品信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AddToCartInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 添加商品到购物车
      tags:
      - cart
  /cart/{id}:
    delete:
      consumes:
      - application/json
      description: 从购物车中删除指定商品
      parameters:
      - description: 购物车项ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除购物车商品
      tags:
      - cart
    put:
      consumes:
      - application/json
      description: 更新购物车中指定商品的数量
      parameters:
      - description: 购物车项ID
        in: path
        name: id
        required: true
        type: integer
      - description: 数量信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCartInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新购物车商品数量
      tags:
      - cart
...
  /favorites:
    get:
      consumes:
      - application/json
      description: 获取当前用户的收藏列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取收藏列表
      tags:
      - favorites
    post:
      consumes:
      - application/json
      description: 添加商品到收藏列表
      parameters:
      - description: 商品信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AddFavoriteInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 添加收藏
      tags:
      - favorites
  /favorites/{productId}:
    delete:
      consumes:
      - application/json
      description: 从收藏列表中删除指定商品
      parameters:
      - description: 产品ID
        in: path
        name: productId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除收藏
      tags:
      - favorites
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"